backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

sqlalchemy==2.0.30

pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1

# Optional: enables higher-quality parsing if OPENAI_API_KEY is set
openai==1.30.1

# Testing
pytest==8.2.2
httpx==0.27.0
